requirements.txt
sqlmodel
sqlalchemy
fastapi
annotated-types==0.7.0
anyio==4.11.0
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
distro==1.9.0
exceptiongroup==1.3.0
Flask==3.1.2
Flask-JWT-Extended==4.7.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
openai==2.2.0
OpenAIAuth==3.0.0
prompt_toolkit==3.0.52
pydantic==2.11.10
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
PySocks==1.7.1
python-dotenv==1.1.1
python-telegram-bot==22.5
regex==2025.9.18
requests==2.32.5
revChatGPT==6.8.6
rich==14.1.0
sniffio==1.3.1
socksio==1.0.0
tiktoken==0.12.0
tls-client==1.0.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.2.14
Werkzeug==3.1.3
